Revert "Lesson 6 and 7 Relation collection with API"
