--- MERN-Stack.postman_collection lesson 6 -7.json
@@ -1,302 +1,0 @@
-{
-	"info": {
-		"_postman_id": "d1e45599-437a-4b9b-aec5-38a7af06b4e0",
-		"name": "MERN-Stack",
-		"schema": "https://schema.getpostman.com/json/collection/v2.1.0/collection.json",
-		"_exporter_id": "20764093"
-	},
-	"item": [
-		{
-			"name": "Get HealthCheck",
-			"request": {
-				"method": "GET",
-				"header": [],
-				"url": {
-					"raw": "http://localhost:3000/healthCheck",
-					"protocol": "http",
-					"host": [
-						"localhost"
-					],
-					"port": "3000",
-					"path": [
-						"healthCheck"
-					]
-				}
-			},
-			"response": []
-		},
-		{
-			"name": "Post Posts",
-			"request": {
-				"method": "POST",
-				"header": [],
-				"body": {
-					"mode": "raw",
-					"raw": "{\r\n    \"title\":\"text me chamath B`\",\r\n    \"content\":\"Hii Chama You Win A\",\r\n    \"firstName\":\"chamthka A\"\r\n}",
-					"options": {
-						"raw": {
-							"language": "json"
-						}
-					}
-				},
-				"url": {
-					"raw": "http://localhost:3000/api/posts/",
-					"protocol": "http",
-					"host": [
-						"localhost"
-					],
-					"port": "3000",
-					"path": [
-						"api",
-						"posts",
-						""
-					]
-				}
-			},
-			"response": []
-		},
-		{
-			"name": "User Register",
-			"request": {
-				"method": "POST",
-				"header": [],
-				"body": {
-					"mode": "raw",
-					"raw": "{\r\n    \"email\":\"test12@gmail.com\",\r\n    \"username\":\"test12345\",\r\n    \"password\":\"test@12345\"\r\n}",
-					"options": {
-						"raw": {
-							"language": "json"
-						}
-					}
-				},
-				"url": {
-					"raw": "http://localhost:3000/api/user/register",
-					"protocol": "http",
-					"host": [
-						"localhost"
-					],
-					"port": "3000",
-					"path": [
-						"api",
-						"user",
-						"register"
-					]
-				}
-			},
-			"response": []
-		},
-		{
-			"name": "User Register Login",
-			"request": {
-				"method": "POST",
-				"header": [],
-				"body": {
-					"mode": "raw",
-					"raw": "{\r\n    \"username\":\"test12345\",\r\n    \"password\":\"test@12345\"\r\n}",
-					"options": {
-						"raw": {
-							"language": "json"
-						}
-					}
-				},
-				"url": {
-					"raw": "http://localhost:3000/api/user/login",
-					"protocol": "http",
-					"host": [
-						"localhost"
-					],
-					"port": "3000",
-					"path": [
-						"api",
-						"user",
-						"login"
-					]
-				}
-			},
-			"response": []
-		},
-		{
-			"name": "Task Category",
-			"request": {
-				"method": "POST",
-				"header": [],
-				"body": {
-					"mode": "raw",
-					"raw": "{\r\n    \"name\":\"Home Task\"\r\n}",
-					"options": {
-						"raw": {
-							"language": "json"
-						}
-					}
-				},
-				"url": {
-					"raw": "http://localhost:3000/api/task/category",
-					"protocol": "http",
-					"host": [
-						"localhost"
-					],
-					"port": "3000",
-					"path": [
-						"api",
-						"task",
-						"category"
-					]
-				}
-			},
-			"response": []
-		},
-		{
-			"name": "Task Add",
-			"request": {
-				"method": "POST",
-				"header": [],
-				"body": {
-					"mode": "raw",
-					"raw": "{\r\n    \"category_id\":\"670b4812dc69350b0df898ef\",\r\n    \"title\":\"Task 1\",\r\n    \"content\":\"task 1 content\"\r\n}",
-					"options": {
-						"raw": {
-							"language": "json"
-						}
-					}
-				},
-				"url": {
-					"raw": "http://localhost:3000/api/task/add",
-					"protocol": "http",
-					"host": [
-						"localhost"
-					],
-					"port": "3000",
-					"path": [
-						"api",
-						"task",
-						"add"
-					]
-				}
-			},
-			"response": []
-		},
-		{
-			"name": "Get Task List",
-			"request": {
-				"method": "GET",
-				"header": [],
-				"url": {
-					"raw": "http://localhost:3000/api/task/list",
-					"protocol": "http",
-					"host": [
-						"localhost"
-					],
-					"port": "3000",
-					"path": [
-						"api",
-						"task",
-						"list"
-					]
-				}
-			},
-			"response": []
-		},
-		{
-			"name": "Get All Posts",
-			"request": {
-				"method": "GET",
-				"header": [],
-				"url": {
-					"raw": "http://localhost:3000/api/posts/",
-					"protocol": "http",
-					"host": [
-						"localhost"
-					],
-					"port": "3000",
-					"path": [
-						"api",
-						"posts",
-						""
-					]
-				}
-			},
-			"response": []
-		},
-		{
-			"name": "Get All Post By Id",
-			"request": {
-				"method": "GET",
-				"header": [],
-				"url": {
-					"raw": "http://localhost:3000/api/post/6706b9f31e2f1fb9bfa44aec",
-					"protocol": "http",
-					"host": [
-						"localhost"
-					],
-					"port": "3000",
-					"path": [
-						"api",
-						"post",
-						"6706b9f31e2f1fb9bfa44aec"
-					]
-				}
-			},
-			"response": []
-		},
-		{
-			"name": "Put API Update Post By Id",
-			"request": {
-				"method": "PUT",
-				"header": [],
-				"body": {
-					"mode": "raw",
-					"raw": "{\r\n    \"title\":\"updated title by chamasa\",\r\n    \"content\":\"updated content by chamasa\"\r\n}",
-					"options": {
-						"raw": {
-							"language": "json"
-						}
-					}
-				},
-				"url": {
-					"raw": "http://localhost:3000/api/post/6706b9f31e2f1fb9bfa44aec",
-					"protocol": "http",
-					"host": [
-						"localhost"
-					],
-					"port": "3000",
-					"path": [
-						"api",
-						"post",
-						"6706b9f31e2f1fb9bfa44aec"
-					]
-				}
-			},
-			"response": []
-		},
-		{
-			"name": "DELETE API DELETE Post By Id",
-			"request": {
-				"method": "DELETE",
-				"header": [],
-				"body": {
-					"mode": "raw",
-					"raw": "{\r\n    \"title\":\"updated title by chamas\",\r\n    \"content\":\"updated content by chamas\"\r\n}",
-					"options": {
-						"raw": {
-							"language": "json"
-						}
-					}
-				},
-				"url": {
-					"raw": "http://localhost:3000/api/post/6706bbfd974d1a11b905acde",
-					"protocol": "http",
-					"host": [
-						"localhost"
-					],
-					"port": "3000",
-					"path": [
-						"api",
-						"post",
-						"6706bbfd974d1a11b905acde"
-					]
-				}
-			},
-			"response": []
-		}
-	]
-}
