+++ package.json
@@ -9,8 +9,8 @@
   "author": "Chamath",
   "license": "ISC",
   "dependencies": {
-    "body-parser": "^1.20.3",
+    "body-parser": "^1.20.2",
     "express": "^4.18.3",
-    "mongoose": "^8.7.0"
+    "mongoose": "^8.2.4"
   }
 }
